Manually edited proxy for timerResolution per 1F:#12008
There was a bad proxy path located in
TestFramework/models/TestFramework/Components/TestSequencer/TestSequencer/TestSequencer.xtuml.

The path to the timerResolution EDT was incorrectly set to:
../../../Datatypes/Datatypes.xtuml, changing it to
../../../SharedDatatypes/SharedDatatypes.xtuml addresses the particular
issue.

It is possible that when a package reference is added, the proxy data
for the types included within the referenced package are not updated.