Remove sequence diagrams
This branch is a work-around for an erroneous BridgePoint schema
embedded in pyxtuml prebuilder and is not intended to be merged into
master.  Removing the sequence diagrams prevents CoreSqlLoader from
using an out-of-bounds array index on MSG_M because the schema used by
pyxtuml expects only three attributes for this class while the BP7
schema contains four.